+++ terraform-v2/CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to the orb will be documented in this file.
 Orbs are immutable, some orb versions with no significant changes are
 not listed
 
+## ovotech/terraform-v2@2.5.1
+- Hide Terraform encryption key in Github comments
+
 ## ovotech/terraform-v2@2.4.22
 - Added Terraform provider Aiven Kafka Users v1.1.1 and v1.1.2 to all executors